requirements.txt
psycopg2-binary>=2.9
autogen-agentchat
autogen-ext[openai]
qdrant-client
transformers
torch
nltk
fastapi
uvicorn
httpx
websockets
redis
rq
pytrec-eval-terrier
